+++ src/main/java/com/geraldsaccount/neuefische_todo/model/undo/UndoableAction.java
@@ -2,7 +2,7 @@ package com.geraldsaccount.neuefische_todo.model.undo;
 
 import java.util.Optional;
 
-import com.geraldsaccount.neuefische_todo.model.Task;
+import com.geraldsaccount.neuefische_todo.model.tasks.Task;
 
 public interface UndoableAction {
 	Optional<Task> undo();
